+++ src/app/d2-vendors/Vendor.tsx
@@ -76,11 +76,10 @@ export default class Vendor extends React.Component<Props> {
       .filter((n) => n.length)
       .join(', ');
 
-    const vendorEngramDrops = getVendorDropsForVendor(vendor.vendorHash, allVendorEngramDrops);
+    const vendorEngramDrops = $featureFlags.vendorEngrams
+      ? getVendorDropsForVendor(vendor.vendorHash, allVendorEngramDrops)
+      : [];
     const dropActive = vendorEngramDrops.some(isVerified380);
-
-    const vendorEngramClass = classNames('fa', 'xyz-engram', { 'xyz-active-throb': dropActive });
-
     const vendorLinkTitle = dropActive ? 'VendorEngramsXyz.Likely380' : 'VendorEngramsXyz.Vote';
 
     return (
@@ -92,7 +91,7 @@ export default class Vendor extends React.Component<Props> {
               powerLevelMatters(basePowerLevel) && (
                 <a target="_blank" rel="noopener" href="https://vendorengrams.xyz/">
                   <img
-                    className={vendorEngramClass}
+                    className={classNames('fa', 'xyz-engram', { 'xyz-active-throb': dropActive })}
                     src={vendorEngramSvg}
                     title={t(vendorLinkTitle)}
                   />